Fixed: Directory Deletion issue when cache_keys setting was abused.

+++ src/Admin/Actions.php
@@ -196,7 +196,12 @@ class Actions {
 				continue;
 			}
 
-			$dir = OMGF_UPLOAD_DIR . '/' . $dir_to_remove;
+			$dir      = OMGF_UPLOAD_DIR . '/' . $dir_to_remove;
+			$realpath = realpath( $dir );
+
+			if ( $dir !== $realpath ) {
+				continue;
+			}
 
 			$this->delete_files( $dir );
